+++ ui/app/components/search-box.js
@@ -20,13 +20,18 @@ export default Component.extend({
   actions: {
     setSearchTerm(e) {
       this.set('_searchTerm', e.target.value);
-      run.debounce(this, updateSearch, this.get('debounce'));
+      run.debounce(this, updateSearch, this.debounce);
+    },
+
+    clear() {
+      this.set('_searchTerm', '');
+      run.debounce(this, updateSearch, this.debounce);
     },
   },
 });
 
 function updateSearch() {
-  const newTerm = this.get('_searchTerm');
+  const newTerm = this._searchTerm;
   this.onChange(newTerm);
   this.set('searchTerm', newTerm);
 }