+++ ui/app/mixins/window-resizable.js
@@ -10,12 +10,12 @@ export default Mixin.create({
 
   setupWindowResize: function() {
     run.scheduleOnce('afterRender', this, () => {
-      this.set('_windowResizeHandler', this.get('windowResizeHandler').bind(this));
-      $(window).on('resize', this.get('_windowResizeHandler'));
+      this.set('_windowResizeHandler', this.windowResizeHandler.bind(this));
+      $(window).on('resize', this._windowResizeHandler);
     });
   }.on('didInsertElement'),
 
   removeWindowResize: function() {
-    $(window).off('resize', this.get('_windowResizeHandler'));
+    $(window).off('resize', this._windowResizeHandler);
   }.on('willDestroyElement'),
 });