+++ ui/mirage/factories/node.js
@@ -5,6 +5,7 @@ import moment from 'moment';
 
 const UUIDS = provide(100, faker.random.uuid.bind(faker.random));
 const NODE_STATUSES = ['initializing', 'ready', 'down'];
+const NODE_CLASSES = provide(7, faker.company.bsBuzz.bind(faker.company));
 const REF_DATE = new Date();
 
 export default Factory.extend({
@@ -12,6 +13,7 @@ export default Factory.extend({
   name: i => `nomad@${HOSTS[i % HOSTS.length]}`,
 
   datacenter: faker.list.random(...DATACENTERS),
+  nodeClass: faker.list.random(...NODE_CLASSES),
   drain: faker.random.boolean,
   status: faker.list.random(...NODE_STATUSES),
   tls_enabled: faker.random.boolean,
@@ -106,9 +108,9 @@ export default Factory.extend({
   }),
 
   afterCreate(node, server) {
-    // Each node has a corresponding client stats resource that's queried via node IP.
+    // Each node has a corresponding client stat resource that's queried via node IP.
     // Create that record, even though it's not a relationship.
-    server.create('client-stats', {
+    server.create('client-stat', {
       id: node.httpAddr,
     });
 
@@ -120,7 +122,7 @@ export default Factory.extend({
       eventIds: events.mapBy('id'),
     });
 
-    server.create('client-stats', {
+    server.create('client-stat', {
       id: node.id,
     });
   },