+++ ui/tests/acceptance/task-logs-test.js
@@ -1,13 +1,18 @@
+import { currentURL } from '@ember/test-helpers';
 import { run } from '@ember/runloop';
-import { test } from 'qunit';
-import moduleForAcceptance from 'nomad-ui/tests/helpers/module-for-acceptance';
+import { module, test } from 'qunit';
+import { setupApplicationTest } from 'ember-qunit';
+import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
 import TaskLogs from 'nomad-ui/tests/pages/allocations/task/logs';
 
 let allocation;
 let task;
 
-moduleForAcceptance('Acceptance | task logs', {
-  beforeEach() {
+module('Acceptance | task logs', function(hooks) {
+  setupApplicationTest(hooks);
+  setupMirage(hooks);
+
+  hooks.beforeEach(async function() {
     server.create('agent');
     server.create('node', 'forceIPv4');
     const job = server.create('job', { createAllocations: false });
@@ -16,20 +21,20 @@ moduleForAcceptance('Acceptance | task logs', {
     task = server.db.taskStates.where({ allocationId: allocation.id })[0];
 
     run.later(run, run.cancelTimers, 1000);
-    TaskLogs.visit({ id: allocation.id, name: task.name });
-  },
-});
+    await TaskLogs.visit({ id: allocation.id, name: task.name });
+  });
 
-test('/allocation/:id/:task_name/logs should have a log component', function(assert) {
-  assert.equal(currentURL(), `/allocations/${allocation.id}/${task.name}/logs`, 'No redirect');
-  assert.ok(TaskLogs.hasTaskLog, 'Task log component found');
-});
+  test('/allocation/:id/:task_name/logs should have a log component', async function(assert) {
+    assert.equal(currentURL(), `/allocations/${allocation.id}/${task.name}/logs`, 'No redirect');
+    assert.ok(TaskLogs.hasTaskLog, 'Task log component found');
+  });
 
-test('the stdout log immediately starts streaming', function(assert) {
-  const node = server.db.nodes.find(allocation.nodeId);
-  const logUrlRegex = new RegExp(`${node.httpAddr}/v1/client/fs/logs/${allocation.id}`);
-  assert.ok(
-    server.pretender.handledRequests.filter(req => logUrlRegex.test(req.url)).length,
-    'Log requests were made'
-  );
+  test('the stdout log immediately starts streaming', async function(assert) {
+    const node = server.db.nodes.find(allocation.nodeId);
+    const logUrlRegex = new RegExp(`${node.httpAddr}/v1/client/fs/logs/${allocation.id}`);
+    assert.ok(
+      server.pretender.handledRequests.filter(req => logUrlRegex.test(req.url)).length,
+      'Log requests were made'
+    );
+  });
 });