+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.9.1 (Unreleased)
+
+FEATURES:
+
+ * vault: Add initial support for Vault namespaces [[GH-5520](https://github.com/hashicorp/nomad/pull/5520)]
+ 
+IMPROVEMENTS:
+
+ * core: Add node name to output of `nomad node status` command in verbose mode [[GH-5224](https://github.com/hashicorp/nomad/pull/5224)]
+
 ## 0.9.0 (April 9, 2019)
 
 __BACKWARDS INCOMPATIBILITIES:__